Split source lifecycle controller
